+++ Runtime/Attributes/ConditionalFieldAttribute.cs
@@ -1,25 +1,28 @@
 using UnityEngine;
 
-public class ConditionalFieldAttribute : PropertyAttribute
+namespace HietakissaUtils.Attributes
 {
-    public readonly string ObjectName;
-    public readonly bool TargetBool = true;
-    public readonly int TargetEnum;
-
-    public ConditionalFieldAttribute(string objectName, bool targetBool)
+    public class ConditionalFieldAttribute : PropertyAttribute
     {
-        ObjectName = objectName;
-        TargetBool = targetBool;
-    }
+        public readonly string ObjectName;
+        public readonly bool TargetBool = true;
+        public readonly int TargetEnum;
 
-    public ConditionalFieldAttribute(string objectName, int targetEnum)
-    {
-        ObjectName = objectName;
-        TargetEnum = targetEnum;
-    }
+        public ConditionalFieldAttribute(string objectName, bool targetBool)
+        {
+            ObjectName = objectName;
+            TargetBool = targetBool;
+        }
 
-    public ConditionalFieldAttribute(string objectName)
-    {
-        ObjectName = objectName;
+        public ConditionalFieldAttribute(string objectName, int targetEnum)
+        {
+            ObjectName = objectName;
+            TargetEnum = targetEnum;
+        }
+
+        public ConditionalFieldAttribute(string objectName)
+        {
+            ObjectName = objectName;
+        }
     }
 }