+++ Runtime/QOL/Camera Shake/CameraShakeSource.cs
@@ -1,5 +1,6 @@
 namespace HietakissaUtils.CameraShake
 {
+    using HietakissaUtils.Attributes;
     using System.Collections.Generic;
     using UnityEngine;