+++ CHANGELOG.md
@@ -3,13 +3,31 @@
 ## WIP
 
 * Complete Console/Command System Refactor
-* Simple Tweening System, HKTween
+* ?Simple Tweening System, HKTween
+
+
+## \[1.3.8] - 10.7.2026
+
+### Changed
+
+* Actually moved attributes to their own HietakissaUtils.Attributes namespace this time instead of just the drawers
+
+
+
+## \[1.3.7] - 10.7.2026
+
+### Changed
+
+* Moved attributes to their own HietakissaUtils.Attributes namespace
+
+### Added
+
+* Favorites tool for HKTool system to save assets and GameObjects as favorites for quick selection
+* HKToolsEditorWindow.ShowPrompt(...) for a ok/cancel prompt for HKTools
 
 
 
 ## \[1.3.6] - 8.11.2025
-
-
 
 ### Changed